refactor: extract lib.rs from binary crate (#41)
Split the codebase into a library crate (skillet_mcp) and a thin binary.

Library (src/lib.rs):
- bm25, git, index, integrity, pack, publish, search, state, validate
- New registry module: init_registry, parse_duration, cache_dir_for_url,
  default_cache_dir (extracted from main.rs)
- All 65 unit tests moved to the library

Binary (src/main.rs):
- CLI parsing (clap), MCP server setup (tower-mcp), transport management
- tools/ and resources/ modules (tower-mcp dependent)
- 8 MCP integration tests

cargo test --lib now works. Doc-tests enabled.
